refactor some errors and test some feature(manual testing). add all DTOs for project(all entities). add @Valid validation for project(all entities). make All can register.

+++ src/main/java/com/example/producttestapi/dto/CartItemDto.java
@@ -7,5 +7,5 @@ import lombok.Setter;
 @Setter
 public class CartItemDto {
     private int quantityToTake;
-    private int pricePerItem;
+    private Double pricePerItem;
 }
